Reworked package
* Split up categories to seperate files
* All exports are now sorted to be alphabetical (excluding `constants`)
* Added some unit tests and CI pipeline
