+++ .travis.yml
@@ -1,0 +1,6 @@
+language: node_js
+node_js:
+  - "8"
+script:
+  - npm run lint
+  - npm run test/coverage && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js
