+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "warframe-item-list",
-  "version": "3.0.0",
+  "version": "4.0.0",
   "description": "A list of all Warframe items that contribute to player mastery rank.",
   "keywords": [
     "warframe"
@@ -23,12 +23,16 @@
   },
   "scripts": {
     "lint": "eslint ./",
-    "prepublishOnly": "npm run lint"
+    "prepublishOnly": "npm run lint && npm run test",
+    "test": "jest",
+    "test/coverage": "jest --coverage"
   },
   "dependencies": {},
   "devDependencies": {
+    "coveralls": "^3.0.0",
     "eslint": "^4.19.1",
     "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-plugin-import": "^2.11.0"
+    "eslint-plugin-import": "^2.11.0",
+    "jest": "^22.4.3"
   }
 }
