+++ src/index.js
@@ -1,6 +1,6 @@
-const { array } = require('./array.js');
-const { objects } = require('./object.js');
+const array = require('./array.js');
 const constants = require('./constants.js');
+const objects = require('./objects.js');
 
 module.exports = {
   version: '22.18.7',
