--- src/object.js
@@ -1,28 +1,0 @@
-const { amps } = require('./data/amps.js');
-const { archwings } = require('./data/archwings.js');
-const { archguns, archmelees } = require('./data/archwingWeapons.js');
-const { sentinels, kubrows, kavats } = require('./data/companions.js');
-const { sentinelWeapons } = require('./data/sentinelWeapons.js');
-const { warframes } = require('./data/warframes.js');
-const { primaries, secondaries, melees } = require('./data/weapons.js');
-const { zaws } = require('./data/zaws.js');
-
-const objects = {
-  AMPS: amps,
-  ARCHWINGS: archwings,
-  ARCHWING_GUNS: archguns,
-  ARCHWING_MELEES: archmelees,
-  COMPANION_SENTINELS: sentinels,
-  COMPANION_KUBROWS: kubrows,
-  COMPANION_KAVATS: kavats,
-  SENTINEL_WEAPONS: sentinelWeapons,
-  WARFRAMES: warframes,
-  WEAPON_PRIMARIES: primaries,
-  WEAPON_SECONDARIES: secondaries,
-  WEAPON_MELEES: melees,
-  ZAWS: zaws,
-};
-
-module.exports = {
-  objects,
-};
